Add name and summary to appdata
The appstream spec alleges that <name> and <summary> are optional if you
have a .desktop file with Name= and Comment=, but this appears not to be
true and other projects do indeed duplicate these fields.

I've got these translations by filling in <_name> and <_summary> in
https://github.com/transmission/transmission/pull/224 and building that
branch (which pulls in the existing translations for the Name= and
Comment= in .desktop). Of course one would prefer this to be done
automatically...